BACKLOG-12783 : refresh correcty detail panel
