+++ src/javascript/Edit/engineTabs/useRegisterEngineTabActions.js
@@ -1,4 +1,3 @@
-import {useEffect} from 'react';
 import {engineTabsPermissionCheckQuery} from '~/Edit/engineTabs/engineTabs.permission.gql-query';
 import {registry} from '@jahia/ui-extender';
 import openEngineTabsAction from '~/Edit/engineTabs/openEngineTabs.action';
@@ -13,6 +12,11 @@ export const useRegisterEngineTabActions = () => {
     const {nodeData, site} = useContentEditorContext();
     const {path, displayName, uuid, mixinTypes, primaryNodeType} = nodeData;
 
+    // SINCE DX 7.5 this fct is introduce, not usable by previous DX version
+    if (!window.authoringApi.getEditTabs) {
+        console.warn('DX version is not able to load GWT engine tabs in content editor');
+    }
+
     // Get tabs from GWT authoring API
     const tabs = window.authoringApi.getEditTabs(
         path,
@@ -23,33 +27,24 @@ export const useRegisterEngineTabActions = () => {
         primaryNodeType.hasOrderableChildNodes
     );
 
-    // SINCE DX 7.5 this fct is introduce, not usable by previous DX version
-    if (!window.authoringApi.getEditTabs) {
-        console.warn('DX version is not able to load GWT engine tabs in content editor');
-    }
-
     const {loading, error, data} = useQuery(engineTabsPermissionCheckQuery(tabs, site));
 
-    useEffect(() => {
-        if (!error && !loading) {
-            if (tabs) {
-                // Permission check query
-                const actionPrefix = 'contentEditorGWTTabAction_';
-                const actionStartPriority = 3;
+    if (!error && !loading) {
+        // Permission check query
+        const actionPrefix = 'contentEditorGWTTabAction_';
+        const actionStartPriority = 3;
 
-                tabs
-                    .forEach((tab, index) => {
-                        if (!registry.get(actionPrefix + tab.id) && (!tab.requiredPermission || data.jcr.nodeByPath[tab.id])) {
-                            registry.addOrReplace('action', actionPrefix + tab.id, openEngineTabsAction, {
-                                buttonLabel: tab.title,
-                                targets: ['AdvancedOptionsActions:' + (index + actionStartPriority)],
-                                tabs: [tab.id]
-                            });
-                        }
+        tabs
+            .forEach((tab, index) => {
+                if (!registry.get(actionPrefix + tab.id) && (!tab.requiredPermission || data.jcr.nodeByPath[tab.id])) {
+                    registry.addOrReplace('action', actionPrefix + tab.id, openEngineTabsAction, {
+                        buttonLabel: tab.title,
+                        targets: ['AdvancedOptionsActions:' + (index + actionStartPriority)],
+                        tabs: [tab.id]
                     });
-            }
-        }
-    }, [data, error, loading]);
+                }
+            });
+    }
 
     return {loading, error};
 };
